@@ -1,69 +1,63 @@
 {
   "name": "documentation",
-  "version": "2.1.2",
+  "version": "0.0.0",
   "private": true,
-  "license": "MIT",
   "scripts": {
     "docusaurus": "docusaurus",
-    "start": "pnpm run download-theme && pnpm run generate-markdown && pnpm run generate-changelog && docusaurus start -h 127.0.0.1",
-    "build": "pnpm run download-theme && pnpm run generate-markdown && pnpm run generate-changelog && docusaurus build",
+    "start": "pnpm build.docs && docusaurus start",
+    "build": "pnpm build.docs && docusaurus build",
+    "build.docs": "tsx --tsconfig ./scripts/tsconfig.json ./scripts/build.ts",
     "swizzle": "docusaurus swizzle",
     "deploy": "docusaurus deploy",
     "clear": "docusaurus clear",
-    "rimraf": "rimraf ./static/webcomponent-examples/*",
     "serve": "docusaurus serve",
     "write-translations": "docusaurus write-translations",
     "write-heading-ids": "docusaurus write-heading-ids",
-    "typecheck": "tsc",
-    "download-theme": "ts-node -P ./scripts/tsconfig.json ./scripts/download-theme.ts",
-    "generate-markdown": "ts-node -P ./scripts/tsconfig.json ./scripts/generate-markdown.ts",
-    "generate-changelog": "ts-node -P ./scripts/tsconfig.json ./scripts/generate-changelog.ts",
-    "generate-previews": "ts-node -P ./scripts/tsconfig.json ./scripts/generate-preview-snippets.ts"
+    "typecheck": "tsc"
   },
   "dependencies": {
-    "@docusaurus/core": "3.3.0",
-    "@docusaurus/plugin-client-redirects": "3.3.0",
-    "@docusaurus/preset-classic": "3.3.0",
-    "@docusaurus/theme-live-codeblock": "3.3.0",
+    "@docusaurus/core": "3.7.0",
+    "@docusaurus/preset-classic": "3.7.0",
     "@mdx-js/react": "^3.0.0",
-    "@siemens/ix": "workspace:*",
-    "@siemens/ix-echarts": "workspace:*",
-    "@siemens/ix-icons": "^2.0.0",
-    "@siemens/ix-react": "workspace:*",
-    "@stackblitz/sdk": "^1.8.1",
-    "angular-test-app": "workspace:*",
-    "clsx": "^1.2.1",
-    "docusaurus-lunr-search": "^3.3.0",
-    "docusaurus-plugin-sass": "^0.2.2",
-    "execa": "^5.1.1",
-    "html-test-app": "workspace:*",
-    "ionic-test-app": "workspace:*",
-    "prism-react-renderer": "^2.1.0",
-    "raw-loader": "^4.0.2",
+    "@siemens/ix-icons": "^2.2.0",
+    "@stackblitz/sdk": "^1.11.0",
+    "prism-react-renderer": "^2.3.0",
     "react": "^18.0.0",
-    "react-anime": "^4.1.1",
-    "react-dom": "^18.0.0",
-    "react-test-app": "workspace:*",
-    "sass": "^1.54.8",
-    "vue-test-app": "workspace:*"
+    "react-dom": "^18.0.0"
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "3.3.0",
-    "@docusaurus/tsconfig": "3.3.0",
-    "@docusaurus/types": "3.3.0",
+    "@docusaurus/module-type-aliases": "3.7.0",
+    "@docusaurus/tsconfig": "3.7.0",
+    "@docusaurus/types": "3.7.0",
+    "@siemens/ix": "workspace:*",
+    "@siemens/ix-react": "workspace:*",
     "@types/fs-extra": "^9.0.13",
+    "@types/mustache": "^4.1.1",
     "@types/node": "^20.16.0",
-    "@types/rimraf": "^3.0.2",
-    "@types/tar": "^6.1.7",
-    "axios": "^1.5.1",
+    "@types/react": "^18.3.10",
+    "@types/rimraf": "^4.0.5",
+    "@types/tar": "^6.1.13",
+    "angular-test-app": "workspace:*",
+    "autoprefixer": "10.4.5",
+    "axios": "^1.5.0",
+    "clsx": "^2.1.0",
+    "docusaurus-plugin-sass": "^0.2.5",
     "figma-plugin": "workspace:*",
-    "listr2": "^7.0.1",
-    "octokit": "^3.2.0",
-    "rimraf": "^3.0.2",
-    "tar": "^6.2.0",
-    "ts-node": "^10.9.1",
-    "typedoc": "^0.24.8",
-    "typescript": "~5.2.2"
+    "fs-extra": "^11.2.0",
+    "html-test-app": "workspace:*",
+    "ionic-test-app": "workspace:*",
+    "listr2": "^8.2.5",
+    "mustache": "^4.2.0",
+    "octokit": "^3",
+    "postcss": "^8.4.29",
+    "react-test-app": "workspace:*",
+    "rimraf": "^6.0.1",
+    "tailwindcss": "^3.4.12",
+    "tar": "^7.4.3",
+    "tsx": "^4.15.4",
+    "typedoc": "^0.27.9",
+    "typescript": "~5.5.2",
+    "vue-test-app": "workspace:*"
   },
   "browserslist": {
     "production": [
