@@ -1,11 +1,4 @@
-/*
- * SPDX-FileCopyrightText: 2024 Siemens AG
- *
- * SPDX-License-Identifier: MIT
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
+import type { SidebarsConfig } from '@docusaurus/plugin-content-docs';
 
 /**
  * Creating a sidebar enables you to:
@@ -17,122 +10,237 @@
 
  Create as many sidebars as you want.
  */
-
-import type { SidebarsConfig } from '@docusaurus/plugin-content-docs';
-
 const sidebars: SidebarsConfig = {
-  mySidebar: [
+  // By default, Docusaurus generates a sidebar from the docs folder structure
+  // tutorialSidebar: [{type: 'autogenerated', dirName: '.'}],
+
+  // But you can create a sidebar manually
+  home: [
+    'home/overview',
     {
-      type: 'doc',
-      id: 'introduction',
-    },
-    {
-      type: 'doc',
-      id: 'faq',
-    },
-    {
-      type: 'doc',
-      id: 'contact-us',
-    },
-    {
-      type: 'doc',
-      id: 'starter-app',
+      type: 'category',
+      label: 'Getting started',
+      items: [
+        'home/getting-started/developers',
+        'home/getting-started/designers',
+        'home/getting-started/starter-app',
+      ],
+      collapsed: false,
     },
     {
       type: 'category',
-      label: "What's happening?",
+      label: 'Migration',
       items: [
-        {
-          type: 'doc',
-          id: 'roadmap',
-        },
-        {
-          type: 'doc',
-          id: 'release-info',
-        },
-        {
-          type: 'doc',
-          id: 'installation/CHANGELOG',
-        },
+        'home/migration/2_0_0/index',
+        'home/migration/core-ui',
+        'home/migration/uxt',
       ],
+      collapsed: true,
     },
     {
       type: 'category',
-      label: 'Designing',
+      label: 'Installation',
       items: [
-        {
-          type: 'doc',
-          id: 'getting-started-for-designers',
-        },
-        {
-          type: 'doc',
-          id: 'design-kit/design-kit',
-        },
+        'home/installation/angular',
+        'home/installation/react',
+        'home/installation/vue',
+        'home/installation/csp',
+        'home/installation/blazor',
       ],
+      collapsed: false,
     },
     {
       type: 'category',
-      label: 'Developing',
+      label: 'Theme',
       items: [
-        {
-          type: 'doc',
-          id: 'getting-started',
-        },
-        {
-          type: 'category',
-          label: 'Installation',
-          items: [
-            {
-              type: 'doc',
-              id: 'installation/installation',
-            },
-            {
-              type: 'doc',
-              id: 'installation/angular',
-            },
-            {
-              type: 'doc',
-              id: 'installation/react',
-            },
-            {
-              type: 'doc',
-              id: 'installation/javascript',
-            },
-            {
-              type: 'doc',
-              id: 'installation/vue',
-            },
-            {
-              type: 'doc',
-              id: 'installation/blazor',
-            },
-            {
-              type: 'doc',
-              id: 'installation/csp',
-            },
-          ],
-        },
+        'home/theming/theme',
+        'home/theming/usage-developers',
+        'home/theming/usage-designers',
       ],
+      collapsed: false,
     },
     {
       type: 'category',
-      label: 'UX Writing',
+      label: 'Releases',
       items: [
-        {
-          type: 'autogenerated',
-          dirName: 'language',
-        },
+        'home/releases/roadmap',
+        'home/releases/release-version',
+        'home/releases/changelog',
       ],
+      collapsed: false,
     },
     {
       type: 'category',
-      label: 'CLI design',
+      label: 'Support',
       items: [
+        'home/support/contact-us',
+        'home/support/faq',
         {
-          type: 'autogenerated',
-          dirName: 'cli',
+          type: 'link',
+          href: 'https://github.com/siemens/ix',
+          label: 'GitHub',
+        },
+        {
+          type: 'link',
+          href: 'https://github.com/siemens/ix/blob/main/CONTRIBUTING.md',
+          label: 'Contribution',
         },
       ],
+      collapsed: false,
+    },
+  ],
+  components: [
+    {
+      type: 'category',
+      label: 'Application frame',
+      // link: { type: 'doc', id: 'components/overview' },
+      // link: { type: 'generated-index' },
+      items: [
+        'components/application/index',
+        'components/application-header/index',
+        'components/application-menu/index',
+        'components/content/index',
+        'components/about-and-legal/index',
+        'components/settings/index',
+      ],
+      collapsed: false,
+    },
+    {
+      type: 'category',
+      label: 'Navigation and hierarchy',
+      items: [
+        'components/basic-navigation/index',
+        'components/breadcrumb/index',
+        'components/group/index',
+        'components/map-navigation/index',
+        'components/pagination/index',
+        'components/tabs/index',
+        'components/tree/index',
+        'components/workflow/index',
+      ],
+      collapsed: false,
+    },
+    {
+      type: 'category',
+      label: 'Containers and layout',
+      items: [
+        //TODO (not exist) 'components/action-card/index',
+        'components/blind/index',
+        'components/card/index',
+        'components/card-list/index',
+        'components/flip/index',
+        'components/event-list/index',
+        'components/drawer/index',
+        'components/layout-auto/index',
+        'components/layout-grid/index',
+        'components/modal/index',
+        'components/panes/index',
+        //TODO (not exist) 'components/push-card/index',
+        'components/tile/index',
+      ],
+      collapsed: false,
+    },
+    {
+      type: 'category',
+      label: 'Forms',
+      items: [
+        'components/forms-field/index',
+        'components/forms-layout/index',
+        'components/forms-validation/index',
+        'components/forms-behavior/index',
+      ],
+      collapsed: false,
+    },
+    {
+      type: 'category',
+      label: 'Input fields and selections',
+      items: [
+        'components/category-filter/index',
+        'components/checkbox/index',
+        'components/custom-field/index',
+        'components/date-dropdown/index',
+        'components/input-date/index',
+        'components/date-picker/index',
+        'components/date-time-picker/index',
+        'components/expanding-search/index',
+        'components/input-number/index',
+        'components/radio/index',
+        'components/select/index',
+        'components/slider/index',
+        'components/input/index',
+        'components/textarea/index',
+        'components/time-picker/index',
+        'components/toggle/index',
+        'components/upload/index',
+      ],
+      collapsed: false,
+    },
+    {
+      type: 'category',
+      label: 'Buttons and actions',
+      items: [
+        'components/avatar/index',
+        'components/button/index',
+        // TODO (button-group is part of button) 'components/button/index',
+        'components/dropdown-button/index',
+        'components/icon-button/index',
+        'components/link-button/index',
+        'components/split-button/index',
+        'components/toggle-button/index',
+        'components/chip/index',
+      ],
+      collapsed: false,
+    },
+    {
+      type: 'category',
+      label: 'System feedback and status',
+      items: [
+        'components/empty-state/index',
+        'components/messagebar/index',
+        'components/pill/index',
+        'components/spinner/index',
+        'components/toast/index',
+        'components/tooltip/index',
+      ],
+      collapsed: false,
+    },
+    {
+      type: 'category',
+      label: 'Data display',
+      items: [
+        'components/grid/index',
+        'components/html-grid/index',
+        'components/key-value/index',
+        'components/key-value-list/index',
+        'components/kpi/index',
+      ],
+      collapsed: false,
+    },
+    {
+      type: 'category',
+      label: 'Charts',
+      items: [
+        'components/overview/index',
+        'components/line-chart/index',
+        'components/bar-chart/index',
+        'components/gauge-chart/index',
+        'components/pie-chart/index',
+        'components/3d/index',
+        'components/special-chart/index',
+      ],
+      collapsed: false,
+    },
+  ],
+
+  guidelines: [
+    {
+      type: 'category',
+      label: 'Mobile experience',
+      items: [
+        'guidelines/mobile/mobile-app-dev',
+        'guidelines/mobile/mobile-ux',
+      ],
+      collapsed: false,
     },
     {
       type: 'category',
@@ -140,64 +248,62 @@ const sidebars: SidebarsConfig = {
       items: [
         {
           type: 'autogenerated',
-          dirName: 'conversational-design',
+          dirName: 'guidelines/conversational-design',
         },
       ],
+      collapsed: false,
     },
     {
       type: 'category',
-      label: 'Mobile experience',
+      label: 'UX Writing',
       items: [
         {
           type: 'autogenerated',
-          dirName: 'mobile',
+          dirName: 'guidelines/language',
         },
       ],
+      collapsed: false,
     },
     {
       type: 'category',
-      label: 'Migration',
+      label: 'CLI',
       items: [
         {
           type: 'autogenerated',
-          dirName: 'migration',
+          dirName: 'guidelines/cli',
         },
       ],
+      collapsed: false,
+    },
+  ],
+  icons: [
+    'icons/icon-library',
+    {
+      type: 'category',
+      label: 'Icon guidelines',
+      // link: { type: 'doc', id: 'components/overview' },
+      // link: { type: 'generated-index' },
+      items: [
+        'icons/guidelines/overview',
+        'icons/guidelines/design-new-icons',
+        'icons/guidelines/icon-design-guidelines',
+      ],
+      collapsed: false,
     },
     {
       type: 'category',
-      label: 'Styles',
-      items: [
-        {
-          type: 'doc',
-          id: 'icon-library/icons',
-        },
-        {
-          type: 'autogenerated',
-          dirName: 'theming',
-        },
-      ],
+      label: 'Development',
+      // link: { type: 'doc', id: 'components/overview' },
+      // link: { type: 'generated-index' },
+      items: ['icons/code/usage', 'icons/code/integration'],
+      collapsed: false,
     },
-    {
-      type: 'category',
-      label: 'Theming',
-      items: [
-        {
-          type: 'autogenerated',
-          dirName: 'guidelines',
-        },
-      ],
-    },
-    {
-      type: 'category',
-      label: 'Components',
-      items: [
-        {
-          type: 'autogenerated',
-          dirName: 'controls',
-        },
-      ],
-    },
+  ],
+  styles: [
+    'styles/colors',
+    'styles/typography',
+    'styles/borders',
+    'styles/shadows',
   ],
 };
 
